Refactoring.  Extracting BedTime module.
