+++ lib/babysitter.ex
@@ -1,9 +1,9 @@
 defmodule BabySitter do
 
   alias BabySitter.ScheduleValidation, as: ScheduleValidation
-  alias BabySitter.BeforeBed, as: BeforeBed
+  alias BabySitter.BeforeBedTime, as: BeforeBedTime
+  alias BabySitter.Bedtime, as: BedTime
 
-  @after_bed_rate 8
   @after_midnight_rate 16
 
   @too_early_message "Babysitter cannot start before 5:00PM"
@@ -19,32 +19,11 @@ defmodule BabySitter do
 
   defp calculate_by_pay_rate(start_time, hours, bed_time) do
     end_time = start_time + hours
-    before_bed_pay = BeforeBed.calculate_before_bed_pay(start_time, end_time, bed_time)
-    after_bed_pay = calculate_bed_time_pay(start_time, end_time, bed_time)
+    before_bed_pay = BeforeBedTime.calculate_before_bed_pay(start_time, end_time, bed_time)
+    after_bed_pay = BedTime.calculate_bed_time_pay(start_time, end_time, bed_time)
     after_midnight_pay = calculate_midnight_pay(end_time)
 
     before_bed_pay + after_bed_pay + after_midnight_pay
-  end
-
-  defp calculate_bed_time_hours(_, end_time, bed_time) when end_time <= bed_time do
-    0
-  end
-
-  defp calculate_bed_time_hours(_, end_time, bed_time) when end_time > 24 do
-    24 - bed_time
-  end
-
-  defp calculate_bed_time_hours(start_time, end_time, bed_time) when end_time <= 24 and start_time < bed_time do
-    end_time - bed_time
-  end
-
-  defp calculate_bed_time_hours(start_time, end_time, bed_time) when start_time >= bed_time do
-    end_time - start_time
-  end
-
-  defp calculate_bed_time_pay(start_time, end_time, bed_time) do
-    after_bed_hours = calculate_bed_time_hours(start_time, end_time, bed_time)
-    after_bed_hours * @after_bed_rate
   end
 
   defp calculate_after_midnight_hours(end_time) when rem(end_time, 24) < 5 do
